check for Preferences, else default to getSystemLookAndFeel()
Ripped of all the copy-pasta for Mac, Win and Linux checking.
Why? I count points for this:
+1 there were copy-paste-code in it
+1 hard coded fonts
+1 too complicated for Look & Feel checking in a reference manager
+1 there is somewhere a fallback to getSystemLookAndFeel, which never get triggered
+1 default was Plastic; it platform-compatibly looks aweful

Tested on Fedora 17 with Gnome 3, nothing special: looks like expected

Regressions in subtle design details which depended on icon sizes or Tahoma fonts are ignored, cause not documented and therefor not testable for me.

Sideeffect:
* JabRef.java - Organized Imports with Eclipse