Remove special fields and refactor entry table configuration.
Special fields is somewhat hacky on JabRef. I do not use them, so
I opted for their removal. Later we can set something similar
by using just the field's name and plugins (instead of relying on
special "icons" that are often useless and take up space).

Another nice change was the complete refactoring of TableColumnsTab.
It does not rely anymore on JCommons and its cleaner.