Refactor class FieldContentParser and create unit tests.
The wrap methods at FieldContentParser were moved to net.sf.jabref.Util (as it
was its solely user).

FieldContentParser.format() is much, much simpler now (and still probably correct).
Instead of the spaguetti of if/else, now it is modelled against a finite state
machine.

FieldContentParser.format() input and output parameters were changed from
StringBuffer to StringBuilder. The current code does not rely on StringBuffer
features (synchronized access), so it was a safe change.

Unit tests were created for FieldContentParser.